requirements.txt
masonite>=4.0,<5.0
masonite-orm>=2.0,<3.0
masonite-inertia
masonite-js-routes
python-slugify
psycopg2-binary